# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Flow API
  version: 1.0.0
  description: Mini SaaS API - Notion/Trello/Slack lite. Generated from JSDoc in src/modules/**/*.routes.js via swagger-jsdoc.
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: http://localhost:3000
    description: Production
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - cookieAuth: []
    bearerAuth: []
paths:
...
  /api/workspaces:
    post: { summary: Create workspace, tags: [Workspaces], security: [{ cookieAuth: [], bearerAuth: [] }], responses: { '201': { description: Created } } }
    get: { summary: List workspaces (cached), tags: [Workspaces], security: [{ cookieAuth: [], bearerAuth: [] }], responses: { '200': { description: List } } }
  /api/workspaces/{id}:
    get: { summary: Get workspace, tags: [Workspaces], parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: Workspace } } }
    put: { summary: Update workspace (owner), tags: [Workspaces], responses: { '200': { description: Updated } } }
    delete: { summary: Delete workspace (owner), tags: [Workspaces], responses: { '200': { description: Deleted } } }
...
  /api/lists/{listId}/cards:
    post: { summary: Create card, tags: [Cards], responses: { '201': { description: Created } } }
    get:
      summary: List cards (paginated, $text + regex fallback search)
      tags: [Cards]
      parameters:
        - { name: listId, in: path, required: true, schema: { type: string } }
        - { name: search, in: query, schema: { type: string }, description: Search title/description via text index + regex fallback }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses: { '200': { description: List } }
...
  /api/workspaces/{workspaceId}/notes:
    post: { summary: Create note, tags: [Notes], responses: { '201': { description: Created } } }
    get:
      summary: List notes (text search)
      tags: [Notes]
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses: { '200': { description: Notes } }
...